# swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IUser": {
			"description": "",
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"id",
				"name"
			],
			"type": "object"
		}
	},
	"info": {
		"description": "blog-engine backend api endpoints documentations",
		"license": {
			"name": "MIT"
		},
		"title": "blog-engine",
		"version": "0.1.14"
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "UserGetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IUser"
							}
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"UserController"
				],
				"parameters": []
			},
			"post": {
				"operationId": "UserPost",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"UserController"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "UserGetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"UserController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			},
			"put": {
				"operationId": "UserPut",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"UserController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			},
			"patch": {
				"operationId": "UserPatch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"UserController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			},
			"delete": {
				"operationId": "UserRemove",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"UserController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		},
		"tsoa_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://swagger.io/api/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify things",
				"read:pets": "read things"
			}
		}
	},
	"host": "dev.localhost",
	"api_key": {
		"type": "apiKey",
		"name": "api_key",
		"in": "header"
	},
	"tsoa_auth": {
		"type": "oauth2",
		"authorizationUrl": "http://swagger.io/api/oauth/dialog",
		"flow": "implicit",
		"scopes": {
			"write:pets": "modify things",
			"read:pets": "read things"
		}
	}
}